git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/concepts@20064 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ runtime/src/comm/mpi/comm-mpi-locales.c
@@ -3,10 +3,10 @@
 #include "chpl-comm-locales.h"
 #include "error.h"
 
-int chpl_comm_default_num_locales(void) {
+int64_t chpl_comm_default_num_locales(void) {
   return chpl_specify_locales_error();
 }
 
 
-void chpl_comm_verify_num_locales(int32_t proposedNumLocales) {
+void chpl_comm_verify_num_locales(int64_t proposedNumLocales) {
 }
