+++ test/extern/sungeun/classes/extern_defs.c
@@ -1,6 +1,6 @@
 #include "extern_defs.h"
 
 void
-foo(C c, int x) {
+foo(C c, int64_t x) {
   c->x = x;
 }
